Move HiveCoercionPolicy and coercers to dedicated package
