+++ presto-hive/src/main/java/com/facebook/presto/hive/HivePageSourceProvider.java
@@ -14,6 +14,7 @@
 package com.facebook.presto.hive;
 
 import com.facebook.presto.hive.HdfsEnvironment.HdfsContext;
+import com.facebook.presto.hive.coercions.HiveCoercionPolicy;
 import com.facebook.presto.spi.ColumnHandle;
 import com.facebook.presto.spi.ConnectorPageSource;
 import com.facebook.presto.spi.ConnectorSession;
@@ -56,6 +57,7 @@ public class HivePageSourceProvider
     private final HdfsEnvironment hdfsEnvironment;
     private final Set<HiveRecordCursorProvider> cursorProviders;
     private final TypeManager typeManager;
+    private final HiveCoercionPolicy coercionPolicy;
 
     private final Set<HivePageSourceFactory> pageSourceFactories;
 
@@ -65,7 +67,8 @@ public class HivePageSourceProvider
             HdfsEnvironment hdfsEnvironment,
             Set<HiveRecordCursorProvider> cursorProviders,
             Set<HivePageSourceFactory> pageSourceFactories,
-            TypeManager typeManager)
+            TypeManager typeManager,
+            HiveCoercionPolicy coercionPolicy)
     {
         requireNonNull(hiveClientConfig, "hiveClientConfig is null");
         this.hiveStorageTimeZone = hiveClientConfig.getDateTimeZone();
@@ -73,6 +76,7 @@ public class HivePageSourceProvider
         this.cursorProviders = ImmutableSet.copyOf(requireNonNull(cursorProviders, "cursorProviders is null"));
         this.pageSourceFactories = ImmutableSet.copyOf(requireNonNull(pageSourceFactories, "pageSourceFactories is null"));
         this.typeManager = requireNonNull(typeManager, "typeManager is null");
+        this.coercionPolicy = requireNonNull(coercionPolicy, "coercionPolicy is null");
     }
 
     @Override
@@ -101,6 +105,7 @@ public class HivePageSourceProvider
                 hiveSplit.getPartitionKeys(),
                 hiveStorageTimeZone,
                 typeManager,
+                coercionPolicy,
                 hiveSplit.getColumnCoercions());
         if (pageSource.isPresent()) {
             return pageSource.get();
@@ -124,6 +129,7 @@ public class HivePageSourceProvider
             List<HivePartitionKey> partitionKeys,
             DateTimeZone hiveStorageTimeZone,
             TypeManager typeManager,
+            HiveCoercionPolicy coercionPolicy,
             Map<Integer, HiveType> columnCoercions)
     {
         List<ColumnMapping> columnMappings = ColumnMapping.buildColumnMappings(partitionKeys, hiveColumns, columnCoercions, path, bucketNumber);
@@ -147,6 +153,7 @@ public class HivePageSourceProvider
                                 columnMappings,
                                 hiveStorageTimeZone,
                                 typeManager,
+                                coercionPolicy,
                                 pageSource.get()));
             }
         }
